Remove runtime chdir and operator requires
chdir is unsafe when running threaded code, and should not be called
by a library at runtime, and it looks like Nodifier.new is called at
runtime by Tree.new.

Additionally, if possible it's best to avoid requiring files at
runtime.  There doesn't seem to be a good reason to avoid requiring
all operators on initialization, so do that, and register the
operators in a hash. This way Nodifier does not need require all
operator files each time an instance is created.  Freeze the hash
of operators for thread safety.

This drops all dependency on facets from Nodifier, as well as
all dependency on Pathname.

One other change is that the operators hash is string keyed instead of
symbol keyed.  This makes it so you don't need to call to_sym in
Nodifier#create, which was probably a denial of service vulnerability
in ruby <2.2.

All of the operator files are modified in this commit, but it's just to
add a line to register the classes.  We may want to remove registrations
for abstract base classes, but I'm not sure which classes are abstract
base classes currently, other than Operator itself.