Optimize some Union operators
This optimizes Union operators that union multiple instances of
the same SourceVocabularyOperator subclass with different values.

This only results in a 10% speed improvement with the test data
(246ms vs 271ms).  However, the speed up will probably be larger for
larger datasets.  The previous unoptimized approach is going to run
separate scans for each of the branches of the union, so speed for
large datasets should be O(n*m), where n is the number of branches
in the union and m is the size of the dataset.  With this approach,
speed should be O(m) for large datasets.

I think it should be possible to extend this approach to optimizing
unions of separate SourceVocaularyOperator subclasses.  That's a
more important optimization, since it should handle many more cases
than this optimization.