Add optimization for union of ICD9, ICD10, Medcode operators
It's possible to optimize a union of these three cases to avoid
an SQL UNION ALL because they all join to the same table.

This adds a ConditionOccurrenceSourceVocabularyOperator subclass
for ICD9, ICD10, Medcode to subclass from.  This class can handle
unions of separate subclasses, by just ORing the conditions in
each of the subclasses.

To make things easier, this moves the scm.source_code =
condition_source_value from WHERE to JOIN ON, since that is the
same for each subclass. With this change the WHERE conditions for
multiple operators can just be joined with OR.

A ConditionOccurrenceSourceVocabularyOperatorUnion subclass is
added to handle a union of multiple
ConditionOccurrenceSourceVocabularyOperators.

Union#optimize will now flatten unions it contains when optimizing.
This fixes Operator#optimized to handle values that aren't
operators, which was a bug that was caught by the flattening of
unions.

Unfortunately, the tests I'm using only show about a 10% speedup
with this.  I actually couldn't find data in the test database
that used an icd10 or medcode.  However, I expect this speedup will
be much greater on databases like Impala that lack indexes.