Initial try at projection support
This adds basic support for projections to ConceptQL.  Instead
of calling query, you call projection.  This returns a
ConceptQL::Query::Projection object, on which you can call
all to get all of the query's rows, or project to get all
rows from the related table.

By default, projections work in cached mode, where it records all
related ids in a hash.  You can pass :cached=>false to operate
in uncached mode, which will allow the same behavior, but
each projection will also result in running the main query again.

To select only specific columns from the related table, you pass
additional arguments to project.

This doesn't add support for projections to the ConceptQL
language.  It shouldn't be difficult to do that, but it's not
actually necessary, as you can run arbitrary projections on
a ConceptQL::Query object without knowing the projections you
want in advance.