Support custom algorithm fetchers
This supports passing an :algorithm_fetcher option to Query for
custom algorithm fetching.  By default, this looks up algorithms
in the :concepts table in the database.  This is a replacement
for the Concept operator, which is being removed.

If the notifier comes across an algorithm node, instead of
instantiating the operator, it fetches the algorithm using the
algorithm_fetcher, then passes the resulting algorithm back to
the tree to be turned into operators, which are then inlined.

This replaces the third argument to Query#initalize with an
options hash, you can now pass a :tree option to specify the
tree.  In addition, for more user friendliness, you can pass
Tree or Nodifier options directly to Query#initialize, and they
will be passed to the Tree or Nodifier that is automatically
created.

While here, remove the overly mocked spec for Query, since
it broke with these changes and doesn't really test what we
want it to test.