Add test framework and validation tests ported from test_conceptql
This adds a simple minitest/spec based testing framework for
testing that conceptql actually produces correct results for
queries, without using any mocking or testing SQL.

The basic idea is that a ConceptQL query is run, and we either
test that the correct criterion_ids are returned for each
expected criterion_type, or if there are too many criterion_ids,
that the number of criterion_ids returned for each type matches
our expectations.

This ports all of the validation tests from test_conceptql to this
new format.  Unfortunately, this results in a couple of broken
tests, where the validation results in test_conceptql don't match
the values returned by conceptql in this test framework.  I've
decided it's best to commit these tests even though they are
failing, as I'm not sure if there bug is in conceptql, a bug in
this test framework, that the validation results in test_conceptql
are wrong, or that the data used by test_conceptql is different
in some way from the test data.