Add tests for most remaining operators, add fixes for Impala
This adds tests for most remaining operators.  It also fixes the
count, date_range, and filter operators to work on Impala.

It makes the from_seer_visits operator handle multiple arguments
as OR instead of AND, as previously when given multiple arguments
it would probably always return no rows.

It implements the visit operator, which apparently didn't work at
all previously.

It removes the SnomedCondition operator, since it was just a copy
of the Snomed operator, though using snomed_condition in a
ConceptQL query will still work.

Some of the new tests required adding data, which I did.  If you've
already setup a test database, you need to run:

  rake test_db_teardown test_db_setup

to rebuild things.