Replace temp-table scope with CTE scope
This replaces temp-table use with CTE use.  This has the advantage
of ripping out all of the code related to prepping queries.

This also adds a test for the recall operator.  With the previous
code, the recall operator test failed, both for the union part
and the except part.

This implementation isn't perfect, mostly to work around the fact
that operators run queries to get columns.  If we can fix things
so that operators always know the columns or can figure them out
without database queries, that will fix the main issue with this
implementation.