Add infrastructure for supporting multiple data models
To use a different data model for a query, you just pass a
:data_model option to Query.new.  This will get passed down to the
Nodifier, which will use it to determine which hash to look in for
operators.

Registering operators now requires specifying the data models
in which to register as arguments.  To use the same operator in
multiple data models, just pass additional arguments.

This doesn't change any existing behavior.  The only currently
supported data model is :omopv4, which is the default for
Nodifier, and is the data model into which all current operators
register.

This removes registering abstract operators that shouldn't be
directly callable by ConceptQL statements, such as operator
superclasses.