Don't issue queries to determine columns
This changes the ConceptQL operators to know what columns they
produce, so that they don't need to query the database to get
that information.

To get this to work, we record the columns for all tables in a
hash, so they can be looked up at runtime.  All of the Operator
subclasses define a query_cols method that returns the columns
in the query.

By default, we assume that query_cols is correct.  However, we
can set the CONCEPTQL_CHECK_COLUMNS environment variable to also
perform the columns query, to check that the specified columns
match the actual columns returned by the database.

There are some operators, such as complement, that call select_it
inside of query, which causes issues.  Work around that problem
by specifying the underlying columns used via :force_columns
dataset option.

This reduces the time to run the tests from 3.1 seconds to 2.0
seconds on PostgreSQL.  I would expect a much more significant
increase on Impala.