Avoid symlink attack in bin/conceptql metadata
It's not safe to write to any file in a known world-writable
location, as an attacker can symlink it to any other file that
you can write to.

One secure way to write to a file in /tmp is to use a mkstemp
or similar function, but I think a simpler solution here is just
to write the file to current directory.  If that isn't advisable,
then metadata should require a file name argument for where
to write the file.