Use original operator in #annotate
When an operator had a label, we were replacing it with a recall
operator.  This is great for optimization, but annotate would return
back a ConceptQL statement that misrepresented labeled nodes as
"recall" operators.

So now we save away the original operator in the Recall operator and
use that original operator when annotating a statement.