Crummy hack on CTEs
When a statement has Recall and some temporal operator, say First
then First creates an "occurrences" CTE and Recall uses another CTE,
say "Heart Attack".  However, the "occurrences" CTE appears first in
the set of CTEs and the "Heart Attack" occurs after it.  BUT the
"occurrences" CTE will reference the "Heart Attack" CTE causing an
error regarding forward references of CTEs.

I tried simply reversing the set of CTEs in the query, and it seems
to solve the problem.