Major refactoring to add CTEs in correct order and simplify things
This significantly changes how the Query internals work.  It removes
Tree, and changes the internal object structure from:

  Query---Tree------Scope
            |         |
        Nodifier   Operator

to:

  Query---Nodifier---Scope
             |
          Operator

With this commit, operators are tied to the nodifier that created
them.  They can access the scope through the nodifier.  Parsing
the ConceptQL expression is now done top down instead of bottom
up, with operators themselves responsible for creating their
upstreams using their nodifier.

Operater.new is overridden so that if an operator has a label,
it returns the related Recall operator.  It also takes the
nodifier object as the first argument.

This required only a few small changes to operator subclasses,
either to add the nodifier as the first argument, or to create
the upstreams.

When creating the upstreams, a stack is used in the scope to
keep track of recall operator usage.  So if a operator with a
label A uses recall operator B inside of it, then A will depend
on B, and the scope will know to add the B CTE before the A CTE.
This avoids having to resort CTEs by trying to parse the SQL
with a regular expression.