Add framework for validations, and upstream validations for operators
This adds a general framework for validations for operator classes.
Each operator class can define validations, and when annotating, it
will validate the operator and upstreams before trying to execute
the query.  If the operator is not valid, it will store errors in
the annotation, instead of counts.  Additionally, the errors and
counts will also be stored in the scope by operator id.

Operator::Invalid has been added, and will automatically be added
for all invalid operator names.  This operator is always invalid.

To test out this framework, validations for upstreams have been
added to all operator classes.  These validations are:

* validate_no_upstreams
* validate_one_upstream
* validate_at_least_one_upstream
* validate_at_most_one_upstream

BinaryOperatorOperator uses exactly two upstreams, but since the
upstreams are stored as options instead of being positional, they
are validated seperately.

Errors and annotations are now cached for each operator.  Operators
should be treated as immutable except for the caches, so this
should be fine.

Other related changes:

* Recall operators no longer store their original operator, they
  can just look it up in the scope.

* Query#scope_annotate has been added, which returns the scope's
  annotations.  This is the method that the API should call to
  get the errors and counts for all parts of the query, keyed
  by operator id.