Add Scope#add_extra_cte for CTEs needed by operators other than recall
These operators should add CTEs after the recall CTEs.  add_extra_cte
will record the CTE that should be added, and add the CTE after the
recall operator CTEs, so that things are added in the proper order.

Modify Occurrence and OneInTwoOut to use add_extra_cte instead
of Dataset#with directly.