Query#annotate returns counts in their own subhash
I've moved count information into it's own subhash within the
"annotation" hash.

I've also included an "invalid" type which is returned when type
information can't be derived.

These efforts were done to make rendering diagrams of ConceptQL a bit
more consistent.  AnnotateGrapher was improperly rendering diagrams
that contained errors.  I'd rather the diagram always gets rendered
so long as the ConceptQL statement is parseable.  If the statement
contains errors, so be it, but the errors should prevent a diagram from
rendering.