Implement TemporalOperator :within option
This was a little more complicated than I expected.  Some temporal
operators only have a end_date or a start_date on the RHS, not
both, so the appropriate :within filter needs to be skipped in
such cases.

For example, in the after operator, the LHS start date needs to
fall after the RHS end date, so there is no point in checking that
the LHS start date falls before the start date less the within
interval.  Similarly for the before, trim_date_end, and
trim_date_start operators.

In order to handle date adjustments in both directions from the
same DateAdjuster, I added an optional second argument to
DateAdjuster#adjust for reversing the quantity.

Note that the current code allow for reverse :within intervals,
allowing you to check for a date inside the RHS start and end dates
by the interval instead of outside.  I'm not sure if this has a
practical use, though I did need to use it in the any_overlap
spec.

I'm not sure in which conditions you want to choose with LHS date
column to use for within.  The spec says after uses the LHS start
date and before uses the LHS end date.  I defaulted
TemporalOperator to use the LHS start date, and then overrode
before and trim_date_end to use the LHS end date.

This does not add tests for all temporal operators, but the
remaining ones only override where_clause and thus should
function similarly to any_overlap, which I did add a test for.