Add support for temporal operator :occurrences option
This is more complicated than adding a single where clause, because
you cannot use window functions in where clauses.  So you need to
use a subselect that selects the window function value in addition
to the existing columns, and filter in the outer select.  You don't
want to return the window function value with the query results, so
you need to select the original columns in the outer query.

Like the :within option, I'm not sure which column from the RHS
should be used when ordering the window function.  Since we
are joining on person_id, I'm guessing that's what the partitioning
should use.  For the order, I defaulted to the RHS start date, but
using the RHS end date if the RHS did not have a start date.