Add concurrent_within operator
This operator takes multiple upstreams and returns records from
each upstream where there is a matching entry for the person
within the time window in all of the other upstreams.

I'm not sure what the behavior should be with a single upstream,
so I defauled to just returning a single upstream as-is, based on
the logic that other upstreams act as filters, and if there are
no other upstreams, then the single dataset should be returned
unfiltered.