Monolithic commit for adding drug info
Introduces Drugish module, which will setup an operator to include
the three additional drug columns we're interested in.  This module
is included in all operators that seem to relate to drugs.

We had to modify the way Operator#modify_query works so that we can
use a class to capture the way a query needs to be modified.  So we now
have a class called QueryModifier which encapsulates the basic API
for query modifications.

Lastly, we modified the way tests work to allow all results to be
checked.  This allows us to ensure that columns are being properly
populated.