Vocab: Fix bugs in SQL generation
Each operator now gets a copy of the name that was used in the ConceptQL
statement to invoke the operator.  That way, operators that generate
metadata for more than one operation are able to know under what name
they've been instantiated.

Now that the Vocabulary operator knows which name was used to invoke it,
it will generate the correct SQL