Windowable: add windowing to Scope/Operators (#123)
We have the need to limit selection of events to a specific window when
we are building a study.

We can now pass in a literal date range, or a table name, to Scope and
have it limit all selection operators to return only events that fall
within the date range or that fall within a date range for a matching
row in the table that is passed in.

This is my first run at implementing this feature.  There are probably
other selection operators that will need this kind of attention, but it
should be as easy as including Windowable in those operators.