Remove Facets-related dependencies (#117)
Facets, though it is AWESOME, is conflicting with ActiveSupport when we
use ConceptQL in Rails.

Rather than replace Facets with ActiveSupport, I've opted to remove most
of the code that needed Facets.  Where I couldn't, I created methods in
ConceptQL::Utils to provide the necessary features we were using from
Facets.