Fix deprecation warnings related to symbol splitting and filter arguments
Symbol splitting will be disabled by default in Sequel 5, so all of
the double and triple underscores need to be removed and replaced
with Sequel objects.  In cases where joining is done, in general
there is no need for explicit qualification, but the implicit
qualification should be correct.

Sequel 5 will no longer allow multiple filter arguments and merge
them together, and will not longer allow Procs not passed as
blocks.  Just use Sequel.expr in the related code, which returns
a single expression.

Also, remove the use of virtual row method blocks, which will
also be removed from Sequel 5.