Use :Bignum instead of Bignum for foreign key types
Bignum constants is deprecated in ruby 2.4, and is actually a
reference to Integer, so it creates a 32-bit integer instead
of a 64-bit integer.  Switch to the :Bignum symbol to continue to
use a 64-bit integer.

Note that I'm not sure this schema is still valid in the
omopv4_plus data model.