Address issue with hanging formatting
Query#format_sql generally calls out to a shell script to generate a
nicely formatted SQL statement.

I've used two shell scripts in the past:
- pg_format
- sqlformat (from sqlparse)

Both have issues with VERY large queries where they seem to get stuck in
infinite loops and never return.

So, I found a Gist on how to timeout a call to the shell and then stuck
the format command in that new method.

Now we'll timeout if it takes longer than 10 seconds to format a query
and simply return the unformatted query.