Fix USE statement execution when testing on impala/hive
Issuing the USE query once via Datase#run only works if only one
connection is ever created.  While that may be true for the current
tests, this will bite us later if we ever have multithreaded access
in the tests or there is an accidental disconnect during testing.

Also, instead of checking for an adapter via a regexp, just check
the database_type.

This code does not handle Sequel's sharding support for
after_connect (where the proc takes a second argument).