Add support for :person_ids scope window option
This also adds support for multiple simultaneous window options,
such as using a date range simultaneously with a person.

The :person_ids option can be any Sequel expression, most commonly
an Integer for a specific person or an array of Integers for
multiple people.

Simplify the Window API somewhat by using #call instead of
windowify, allowing the user of plain lambdas for simpler cases.
The :person_ids scope uses a lambda, and the DateLiteral class
is removed with the code moved to a plain lambda.  The None
class is no longer needs as if no windows are used, an empty
array is used.