Change occurrence at_least/within handling to return all matching occurrences
When at_least/within is not used, occurrence only returns the nth
occurrence in the dataset (at most one row per person in the output).

When at_least/within is used, occurrence's behavior changes so that
all matching rows in the dataset that are the nth occurrence of any
prior events in the dataset are returned, subject to the time
conditions imposed by the at_least/within constraints.

Because the behavior is so different, have query call either
query_simple for the regular case, and query_complex for the
at_least within case.

Add some additional tests, but without the matching result files,
so these tests will also fail.