Optimize union operator by avoiding union in certain cases
If all datasets being unioned are basically the same except for
the WHERE clause, just modify the WHERE clause of the first
dataset to OR in conditions in the other datasets.

Note that this looks only one level down, it won't optimize cases
where there are multiple layers of subqueries.  I'm not sure if
ConceptQL generates such cases, but this is at least good enough
for tests such as union/count_1.