Don't partition fix the person_id column on Impala
The :person_id column identifier should never be constant (I'm
guessing), so this should be safe.  This appears to fix the
window_id different numbering issues on Imapla.  This only
changes behavior if the CONCEPTQL_IN_TEST_MODE environment
variable is set, so it shouldn't affect production, but in
production, it doesn't matter which particular window ids are
used (just that there is a unique one per row).