Move sequel_impala semi_join_first code into ConceptQL
This way it integrates with ConceptQL's query handling so that
ConceptQL is aware of all SQL being executed.  This keeps the
ability to use either subqueries or CTEs for the LEFT SEMI JOIN
expression, with temp tables for the CTEs if ConceptQL is
configured that way.

To implement this, I just moved more code into the rdbms object,
since rdbms.join_options will not handle it any more.

With CONCEPTQL_SEMI_JOIN_FIRST=table there is a failure in at
lease icd9/results_uuid3, which is a union of two identical
icd9 operators with the uuid option.  It duplicates the results
in that case, with all fields being identical except for the
uuid field, so I'm guessing something (maybe the uuid option
logic) needs to updated to support this.  No added failures with
CONCEPTQL_SEMI_JOIN_FIRST=true (which uses subqueries) in the
operators I tested.