Add tests for the From operator
The From operator needs to be tested separately because it cannot
operator on the standard tables in either the gdm or omopv4_plus
data models.  To test it, a temp table has to be created first.

During the testing process, I noticed that the Person operator
was not windowable, which seems like a bug as it breaks ConceptQL
statements such as:

  [:window, [:person], {person_ids: [1]}]

Also, to allow the From operator to be able to use qualified tables
passed in via JSON, I reinstituted __ splitting of strings for the
table.  This is necessary at least for the tests to make sure the
From operator can access tables that were created in a separate,
temporary Impala database/PostgreSQL schema.