Fix breakage caused CONCEPTQL_SORT_TEMP_TABLES=true
b67e839e1ee1ca5aa039f93c9d0622f28f35a434 added the use of :sort_by
option if the CONCEPTQL_SORT_TEMP_TABLES env variable was set.  It
had create_options take a scope argument, but it did not pass the
ConceptQL::Scope object in, it passed a Sequel::Dataset instance in,
and the dataset it passed in was the dataset that was retrieving the
results, not the dataset that was used to build the temporary table.
This is because create_options(self) was used in a define_method
block inside one of the scope methods.  Fix this to actually pass
the scope in, by storing the scope in a local variable.

That change exposed another bug, which is that checking against the
scope's query_columns is not correct in all cases.  At least for the
failing query in question, which uses a VALUES query, we should use
the aliases in the first row of VALUES. To handle this, make
create_options take both the scope and the dataset that will be
used to create the temporary table, and explicitly check for the
:values option.  This code may need to be expanded in the future to
handle more query types.