Vocabulary Overhaul
In order to implement #253, we needed to turn our current dynamic
vocabularies system on its head.

Originally we had a single class that would impersonate all the
dynamically generated vocabularies.  Now we actually create a proper
class for each dynamically generated library and even create OMOP and
GDM-based variants.

Now that each vocabulary is properly represented as a class, we can
modify individual classes to behave as we'd like, so we should be able
to get dynamically generated operators to include lab values much more
easily now.

As with many large overhauls, I ran into other issues that I addressed
along the way:

- Validations we're working correctly and I've addressed an issue where
    identical warnings could stack up within the same operator
- ICD10 is now properly a multi-vocabulary operator
- Removed all hard-coded vocabulary-related operators except for Read
- Put existing behaviors under ConceptQL::Behaviors
- Reworked how requiring operators works