WIP: better validations for operators
In removing _more_ code for columns, I snagged another thread
that I don't have time to keep pulling on.

We now through exceptions when asking a statement for it's SQL
rather than returning the message "SQL unavailable..."

This is creating extra errors in a couple of tests that were otherwise
passing before, mostly in annotations when they go to get counts

We really need to ensure that we don't run annotations against operators
that aren't valid.  This should be quite doable.

Problem is that I don't have time to do it now.

So, in summary:
- Strip out more column-related code
- Ensure validation occurrs before SQL is generated