Finish up work on OHDSI Vocabs
Looks like my changes were cheating on the tests.

There was a lot of work left to be done, but it wasn't apparent because I had
both sets of vocabulary tables available to the tests at all times.

In making this set of updates, I decided that the DataModel class was a better
place to put all the methods we call, rather than Lexicon.  In fact, I
eliminated the use of Lexicon inside of operators, and perhaps we can remove
Lexicon entirely in the future.

I also took the liberty of removing a lot of tests for OMOPv4+